requirements.txt
scapy
python-nmap
requests
matplotlib
pillow
pyshark
pyqt5
psutil
scikit-learn
tensorflow
plotly
sqlalchemy
cryptography
flask
fastapi
asyncio
schedule
slack_sdk
pandas

